lots of stuff, I don't remember

+++ JBotSim/src/controllers/OldProgrammedForager.java
@@ -126,7 +126,7 @@ public class OldProgrammedForager extends Controller {
 	}
 
 	@Override
-	public void controlStep(int time) {
+	public void controlStep(double time) {
 
 		if (nextAction == Action.NA) {
 			nextAction = chooseAction();
