+++ JBotSim/src/controllers/ProgrammedAntiForager.java
@@ -115,7 +115,7 @@ public class ProgrammedAntiForager extends Controller {
 	}
 
 	@Override
-	public void controlStep(int time) {
+	public void controlStep(double time) {
 
 		if (nextAction == Action.NA) {
 			nextAction = chooseAction();
