+++ JBotSim/src/controllers/ProgrammedForager.java
@@ -122,7 +122,7 @@ public class ProgrammedForager extends Controller {
 	}
 
 	@Override
-	public void controlStep(int time) {
+	public void controlStep(double time) {
 
 		if (nextAction == Action.NA) {
 			nextAction = chooseAction();
